Add git commit message rephrasing feature
This commit adds a new feature to rephrase a sentence as a git commit message.
It adds a new command line argument `--git` or `-g` to enable this feature.
It also adds a prompt to the command line to enter the sentence to be rephrased.
Finally, it adds a new event listener to process the input sentence and output the rephrased git commit message.