This is a diff for a code change in the `Loz` class of an NLP library.
The diff contains changes to the `completeUserPrompt` method, which is used to generate a commit message based on the given prompt and diff.

The main changes are:

* Adding a new parameter `max_tokens` to the `OpenAI.Chat.ChatCompletionCreateParams` object, which specifies the maximum number of
tokens that should be generated by the model.
* Changing the default value of the `temperature` parameter from 0 to 1. This will allow the model to generate more diverse and creative
responses.
* Adding a new method `runOpenAIChatCompletion` to the `Loz` class, which takes in a `LLMSettings` object as an argument and returns a
string containing the generated commit message.
* Updating the `completeUserPrompt` method to use the new `runOpenAIChatCompletion` method instead of directly calling the
`OpenAI.Chat.ChatCompletionCreateParams` constructor.

It is worth noting that this diff only contains changes related to the `Loz` class, and does not include any changes to other parts of
the codebase.