The commit message generated by the code is:
* A short summary of the changes made in the commit, which includes the commit message passed as a parameter to the `commit` method.
* The name of the model used for generation, specified as a parameter to the `Loz` constructor.

Here's an example of how this code could be used:
```
const params = {
  model: gpt-3.5-turbo
};

const loz = new Loz(params);

loz.commit(Commit message, async () => {
  // Code to commit changes goes here
});
```
In this example, the `Loz` constructor is called with the `model` parameter set to `gpt-3.5-turbo`. When the `commit` method is called with a commit message, the generated commit message will include the model name in addition to the original commit message passed as a parameter.

Generated by codellama